feat: 테스트를 위한 커맨드와 예외를 분리 합니다.
제대로 예외가 처리되는지 테스트해보기 위해 분리합니다.
